% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{github_GET}
\alias{github_GET}
\title{GET request to GitHub}
\usage{
github_GET(path, ..., pat)
}
\arguments{
\item{path}{API path}

\item{...}{extra arguments passed to httr::GET}

\item{pat}{GitHub personal access token. If missing, uses
\code{github_pat()} from devtools.}
}
\description{
This function is taken directly from devtools (under GPL-3)
}

