% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{snippet_prepare}
\alias{snippet_prepare}
\title{Prepare a snippet for a .snippets file.}
\usage{
snippet_prepare(text)
}
\arguments{
\item{text}{snippet text}
}
\description{
Prepare a snippet for a .snippets file.
}

