% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippet_remove.R
\name{snippet_remove}
\alias{snippet_remove}
\title{Remove a snippet from a .snippets file.}
\usage{
snippet_remove(name, language = "r", path = snippets_path(language),
  error = TRUE)
}
\arguments{
\item{name}{of the snippet to remove}

\item{language}{language of the snippet. Choices are "r", "c_cpp", "html",
"css", "java", "javascript", "python", and "sql". Ignored if \code{path}
is specified.}

\item{path}{path to .snippets file}

\item{error}{whether to raise an error if the snippet is not found}
}
\description{
Remove a snippet from a .snippets file, raising an error if the
snippet is not found in the file.
}
\details{
Note that a .snippets file for a language may not yet exist. If
it does not exist, go to Preferences->Code->Edit Snippets in RStudio,
make a trivial change to the language of choice (for example, a new
empty line), and click Save. This will create the language file in
~/.R/snippets/<language>.snippets with defaults included.
}
\examples{
\dontrun{

# add a .r snippet
snippet_add("dplyr", "${1:object} <- ${1:object} \%>\%\\n${2:function}")

# remove the snippet
snippet_remove("dplyr")

}
}

