% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippet_add.R
\name{snippets_add}
\alias{snippets_add}
\title{Add a list of snippets to a .snippets file.}
\usage{
snippets_add(snippets, language = "r", path = snippets_path(language))
}
\arguments{
\item{snippets}{list of snippets to add}

\item{language}{language of the snippet. Choices are "r", "c_cpp", "html",
"css", "java", "javascript", "python", and "sql". Ignored if \code{path}
is specified.}

\item{path}{path to .snippets file}
}
\description{
Add multiple snippets, in the form of a named list, to a .snippets file.
}
\details{
Note that a .snippets file for a language may not yet exist. If
it does not exist, go to Preferences->Code->Edit Snippets in RStudio,
make a trivial change to the language of choice (for example, a new
empty line), and click Save. This will create the language file in
~/.R/snippets/<language>.snippets with defaults included.
}
\examples{
\dontrun{

# add a .r snippet
snippets_add(list(dplyr = "${1:object} <- ${1:object} \%>\%\\n${2:function}"))

# add an HTML snippet
snp <- "<span style = 'font-size: 10px'>Copyright 2015</span>"
snippet_add(list("copyright", snp, language = "html"))

}
}

