% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippets_get.R
\name{snippets_get}
\alias{snippets_get}
\title{Retrieve a current snippet}
\usage{
snippets_get(name, language = "r", path = snippets_path(language))
}
\arguments{
\item{name}{snippet name: if missing, returns a list of all snippets}

\item{language}{language of the snippet. Choices are "r", "c_cpp", "html",
"css", "java", "javascript", "python", and "sql". Ignored if \code{path}
is specified.}

\item{path}{path to .snippets file}
}
\description{
Retrieve a current snippet
}
\details{
Note that a .snippets file for a language may not yet exist. If
it does not exist, go to Preferences->Code->Edit Snippets in RStudio,
make a trivial change to the language of choice (for example, a new
empty line), and click Save. This will create the language file in
~/.R/snippets/<language>.snippets with defaults included.
}
\examples{
\dontrun{
snippet_get("fun")

# assuming a c_cpp .snippets file exists:
snippet_get("cls", language = "c_cpp")

# get a list of all snippets
snippet_get()

}
}

