% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippets_install.R
\name{snippets_install_github}
\alias{snippets_install_github}
\title{Install one or more snippets from a GitHub repository}
\usage{
snippets_install_github(repo, name = NULL, language = NULL,
  directory = SNIPPET_DIRECTORY)
}
\arguments{
\item{repo}{Username and repository name, e.g. "dgrtwo/snippets"}

\item{name}{name of snippet to install, which must be given along with a language.
If NULL, installs all snippets from the language.}

\item{language}{Language from repo to install. If NULL, installs all snippets}

\item{directory}{Directory to install snippets to}
}
\description{
Install one or more snippets from a GitHub repository. These should be in
the form of one or more <language>.snippets files in the top-level directory.
.snippets files that are not
}
\details{
The \code{directory} argument is available mainly for testing purposes.
}
\examples{
\dontrun{

snippets_install_github("dgrtwo/snippets")

# install just R or C/C++
snippets_install_github("dgrtwo/snippets", language = "r")
snippets_install_github("dgrtwo/snippets", language = "c_cpp")

# install just a single snippet
snippets_install_github("dgrtwo/snippets", language = "r", name = "S3")

}
}

