% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippets_install.R
\name{snippets_install_url}
\alias{snippets_install_url}
\title{Install one or more snippets from a URL.}
\usage{
snippets_install_url(url, name = NULL, language = "r",
  path = snippets_path(language))
}
\arguments{
\item{url}{URL to install from}

\item{name}{name of snippet to install. If NULL, installs all snippets from
the URL.}

\item{language}{language of the snippet. Choices are "r", "c_cpp", "html",
"css", "java", "javascript", "python", and "sql". Ignored if \code{path}
is specified.}

\item{path}{path to .snippets file}
}
\description{
Install one or more snippets by downloading them from a URL.
}
\details{
Note that a .snippets file for a language may not yet exist. If
it does not exist, go to Preferences->Code->Edit Snippets in RStudio,
make a trivial change to the language of choice (for example, a new
empty line), and click Save. This will create the language file in
~/.R/snippets/<language>.snippets with defaults included.
}

