% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippets_read.R
\name{snippets_parse}
\alias{snippets_parse}
\title{Parse a character vector of snippets.}
\usage{
snippets_parse(txt)
}
\arguments{
\item{txt}{character vector of snippets}
}
\description{
Parse a character vector of snippets.
}

