% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{snippets_path}
\alias{snippets_path}
\title{Find the path to a .snippets file.}
\usage{
snippets_path(language = "r")
}
\arguments{
\item{language}{language of snippet file to find}
}
\description{
Find the path to the snippet file of the given language on this computer
(e.g. so that it can be read from or replaced)
}

