% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippets_read.R
\name{snippets_read}
\alias{snippets_read}
\title{Parse the current snippets file into a list of snippets.}
\usage{
snippets_read(language = "r", path = snippets_path(language))
}
\arguments{
\item{language}{language of the snippet. Choices are "r", "c_cpp", "html",
"css", "java", "javascript", "python", and "sql". Ignored if \code{path}
is specified.}

\item{path}{path to .snippets file}
}
\value{
A named list of snippets, each of which is a single string
that includes tabs at the start of each line.
}
\description{
Parse the current snippets file into a list of snippets.
}
\details{
Note that a .snippets file for a language may not yet exist. If
it does not exist, go to Preferences->Code->Edit Snippets in RStudio,
make a trivial change to the language of choice (for example, a new
empty line), and click Save. This will create the language file in
~/.R/snippets/<language>.snippets with defaults included.
}

