% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snippets_write.R
\name{snippets_write}
\alias{snippets_write}
\title{Write a list of snippets to a file.}
\usage{
snippets_write(snippets, path)
}
\arguments{
\item{snippets}{named list of snippets}

\item{path}{path to the file to write to}
}
\description{
This overwrites the file if it exists. To append/edit a snippet, use
\code{\link{snippet_add}} instead.
}

